applied new format for purge products
